// src/components/FlexibleTable/types.ts
type BaseColumn = {
    enableEditing?: boolean;
    enum?: Record<string | number, string>;
    request?: () => Promise<Option[]>;
    tooltip?: string;
    validate?: (value?: unknown) => boolean | string;
    bordered?: boolean;
};

export type TextColumn = {
    type?: 'ShortText';
} & ColumnFieldProps<FieldTextProps>;

export type LongTextColumn = {
    type?: 'LongText';
} & ColumnFieldProps<FieldTextProps>;

export type NumberColumn = {
    type?: 'Number';
} & ColumnFieldProps<FieldNumberProps>;

export type PhoneColumn = {
    type?: 'Phone';
} & ColumnFieldProps<FieldPhoneProps>;

export type EmailColumn = {
    type?: 'Email';
} & ColumnFieldProps<FieldTextProps>;

export type SelectColumn = {
    type?: 'SingleSelection';
    enum: Record<string | number, string>;
    request?: () => Promise<Option[]>;
} & ColumnFieldProps<FieldSelectProps>;

export type MultiSelectColumn = {
    type?: 'MultipleSelection';
    enum: Record<string | number, string>;
    request?: () => Promise<Option[]>;
} & ColumnFieldProps<FieldSelectProps>;

export type DateColumn = {
    type?: 'Date';
} & ColumnFieldProps<FieldDatePickerProps>;

export type LinkColumn = {
    type?: 'Link';
} & ColumnFieldProps<FieldTextProps>;

export type PriorityColumn = {
    type?: 'Priority';
} & ColumnFieldProps<FieldNumberProps>;

export type AssigneeColumn = {
    type?: 'Assignee';
} & ColumnFieldProps<FieldSelectProps>;

export type MultipleAssigneeColumn = {
    type?: 'MultipleAssignee';
} & ColumnFieldProps<FieldSelectProps>;

export type TimeColumn = {
    type?: 'Time';
} & ColumnFieldProps<FieldTimePickerProps>;

export type RichTextColumn = {
    type?: 'RichText';
} & ColumnFieldProps<FieldTextProps>;

export type CountryColumn = {
    type?: 'Country';
} & ColumnFieldProps<FieldCountryProps>;

export type DatetimeColumn = {
    type?: 'Datetime';
} & ColumnFieldProps<FieldDateTimePickerProps>;

export type OriginColumn = {
    type?: 'Origin';
} & ColumnFieldProps<FieldSelectProps>;

export type AttachmentColumn = {
    type?: 'Attachment';
} & ColumnFieldProps<FieldUploadProps>;
export type NotesColumn = {
    type?: 'Notes';
} & ColumnFieldProps<NotesProps>;

export type CurrencyColumn = {
    type?: 'Currency';
} & ColumnFieldProps<FieldCurrencyProps>;

export type CheckboxColumn = {
    type?: 'Checkbox';
} & ColumnFieldProps<FieldCheckboxProps>;

export type ColumnType = BaseColumn &
    (
        | TextColumn
        | NumberColumn
        | SelectColumn
        | MultiSelectColumn
        | DateColumn
        | LongTextColumn
        | LinkColumn
        | PriorityColumn
        | AssigneeColumn
        | MultipleAssigneeColumn
        | TimeColumn
        | PhoneColumn
        | EmailColumn
        | RichTextColumn
        | CountryColumn
        | DatetimeColumn
        | OriginColumn
        | AttachmentColumn
        | NotesColumn
        | CurrencyColumn
        | CheckboxColumn
    );

export type FieldType = ColumnType['type'];

type AssigneeValue = {
    _id: string;
    display_name: string;
};

type PhoneValue = {
    country_code?: CountryCode;
    phone?: string;
    country_calling_code?: string;
    calling_code_with_number?: string;
    national_number?: string;
};

export type OriginValue = {
    type: 'customized' | 'channel';
    data: {
        id?: string;
        type?: string;
        name: string;
    };
};

export type AttachmentValue = {
    type: 'file' | 'url';
    data: {
        url: string;
        name: string;
        key?: string;
        extension?: string;
        uploader?: string;
        user_id?: string;
        sizeInBytes?: number;
        uploadDate?: string;
    };
    extra?: Attachment;
};

export type NotesValue = {
    _id: string;
    author_id: string;
    author_name: string;
    message: string;
    created_at: Date;
    updated_at: Date;
};

export type CurrencyValue = {
    currency_code?: string;
    amounts?: number;
};

export type ColumnValue =
    | string
    | number
    | boolean
    | Date
    | (string | number | Date | undefined | null)[]
    | null
    | AssigneeValue
    | undefined
    | PhoneValue
    | OriginValue
    | AttachmentValue[]
    | NotesValue[]
    | CurrencyValue;

export type Columns<D extends { id: string; _id?: string }> = (Omit<AccessorKeyColumnDef<D, ColumnValue>, 'cell'> &
    ColumnType & {
        cell?: string | ((colProps: CellContext<D, ColumnValue> & { isHover?: boolean; isRowSelected?: boolean }) => any);
    })[];
// ...
export type RequestParameters = {
    sorters?: SortingState;
    filters?: ColumnFiltersState;
    pagination?: PaginationState;
    globalFilter?: string;
};
// ...
interface UpdateCache<D> {
    operation: 'Update';
    record: D;
}

interface DeleteCache {
    operation: 'Delete';
    rowId: string | string[];
}

export type CacheMutationPayload<D> = UpdateCache<D> | DeleteCache;

export interface FlexibleTableRef<D> {
    addNewRecord: () => void;
    refresh: () => void;
    resetPagination: () => void;
    reset: () => void;
    getSelectedRows: any;
    resetRowSelection: () => void;
    removeRowSelection: (rowIds: string[]) => void;
    isSomeSelected: () => boolean;
    getFilterState: () => ColumnFiltersState | undefined;
    resetFilterState: () => void;
    cacheMutation: (payload: CacheMutationPayload<D>) => void;
    setColumnFilters: (filters: ColumnFiltersState) => void;
    getRowModel: () => RowModel<D>;
    handleDataUpdate: (updateData: any) => void;
    handleDataDelete: (rowId: string | string[]) => Promise<boolean | void>;
    updateContainerRect: () => void;
}